naming of google network was always default. This is now named after the cluster name (vpc-xxx)
