+++ upup/pkg/fi/cloudup/gce/instancegroups.go
@@ -17,9 +17,7 @@ limitations under the License.
 package gce
 
 import (
-	"encoding/base32"
 	"fmt"
-	"hash/fnv"
 	"strings"
 
 	context "golang.org/x/net/context"
@@ -216,32 +214,6 @@ func NameForInstanceGroupManager(c *kops.Cluster, ig *kops.InstanceGroup, zone s
 	name := SafeObjectName(shortZone+"."+ig.ObjectMeta.Name, c.ObjectMeta.Name)
 	name = LimitedLengthName(name, 63)
 	return name
-}
-
-// LimitedLengthName returns a string subject to a maximum length
-func LimitedLengthName(s string, n int) string {
-	// We only use the hash if we need to
-	if len(s) <= n {
-		return s
-	}
-
-	h := fnv.New32a()
-	if _, err := h.Write([]byte(s)); err != nil {
-		klog.Fatalf("error hashing values: %v", err)
-	}
-	hashString := base32.HexEncoding.EncodeToString(h.Sum(nil))
-	hashString = strings.ToLower(hashString)
-	if len(hashString) > 6 {
-		hashString = hashString[:6]
-	}
-
-	maxBaseLength := n - len(hashString) - 1
-	if len(s) > maxBaseLength {
-		s = s[:maxBaseLength]
-	}
-	s = s + "-" + hashString
-
-	return s
 }
 
 // matchInstanceGroup filters a list of instancegroups for recognized cloud groups
